MDED-58 String+Attributed, Sequence+Joined перенесены в отдельный файл
